Fix fit typing errors
