+++ examples/detoxification_bias/requirements.txt
@@ -1,13 +1,13 @@
+colorama==0.4.4
+ftfy
+matplotlib
+nltk==3.2.5
 pandas
 scikit-learn
-matplotlib
 seaborn
-ftfy
-tweet-preprocessor
-transformers==3.4.0
-torch==1.7.1
-nltk==3.2.5
-colorama==0.4.4
-torchtext==0.3.1
 tensorboardX
 tokenizers==0.9.2
+torch==1.7.1
+torchtext==0.3.1
+transformers==3.4.0
+tweet-preprocessor
